comparison of integers of different signs

+++ src/CryptoNoteCore/Currency.cpp
@@ -541,7 +541,7 @@ namespace CryptoNote {
 		int64_t solveTime(0);
 		uint64_t difficulty(0), next_difficulty(0);
 		// Loop through N most recent blocks.
-		for (int64_t i = 1; i <= N; i++) {
+		for (int64_t i = 1; i <= (int64_t)N; i++) {
 			solveTime = static_cast<int64_t>(timestamps[i]) - static_cast<int64_t>(timestamps[i - 1]);
 			solveTime = std::min<int64_t>((T * 7), std::max<int64_t>(solveTime, (-6 * T)));
 			difficulty = cumulativeDifficulties[i] - cumulativeDifficulties[i - 1];
